Ai_Resume_detection/requirements.txt
Django==5.2.4
PyPDF2
python-docx
spacy
